Fix integer overflow/undefined behavior in NSIS parser
Fix integer overflow in the NSIS parser

Cast int32_t to uint32_t for comparison with uint32_t, to prevent
integer overflow, as well as signed/unsigned compare warning.

Fixes: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=44493

Also address some other undefined behavior warnings:
* mpool.c: Fixed pointer overflow errors uncovered by UndefinedBehaviorSanitizer.
* matcher-ac.c: Test length to avoid passing NULLs to memcmp.

+++ libclamav/nsis/nulsft.c
@@ -484,7 +484,7 @@ static int nsis_headers(struct nsis_st *n, cli_ctx *ctx)
             pos += 4;
             buf += 4;
         }
-        if ((pos += 4 + nextsz) > n->asz) {
+        if ((pos += (4 + ((uint32_t)nextsz))) > n->asz) {
             n->solid = 1;
             break;
         }
